Add ShapeComponent and move some ComponentPool init code to constructor
I moved the resizing of the sparse and mirror arrays from the init to
constructor for the ComponentPool because sometimes I forget to call the
init function and that leads to subscript out of range errors. Now I
won't get that error if I forget, and at worst I get some performance
hiccups as the game warms up but I can fix that by being more diligent
about calling the init function.